remove getParameterByPara to match.js
